consolidate menu examples
