--- apps/examples/src/examples/custom-actions-menu/README.md
@@ -1,12 +1,0 @@
----
-title: Custom actions menu
-component: ./CustomActionsMenuExample.tsx
-category: ui
-keywords: [drop, down, grouping, rotating, layout, components]
----
-
-You can customize tldraw's actions menu.
-
----
-
-The actions menu is a dropdown menu that can be found in the top-left of the tldraw component, or just above the toolbar on smaller screens. It contains actions related to editing shapes such as grouping, rotating or changing shape order. The actions menu can be customized by providing a `ActionsMenu` component to the `Tldraw` component's `components` prop. If you provide `null`, then that component will be hidden.
