--- apps/examples/src/examples/custom-context-menu/README.md
@@ -1,12 +1,0 @@
----
-title: Custom context menu
-component: ./CustomContextMenuExample.tsx
-category: ui
-keywords: [components, ui]
----
-
-You can customize tldraw's context menu.
-
----
-
-Create some shapes, select them and right click the selection to see the custom context menu. It can be customized by providing a `ContextMenu` component to the `Tldraw` component's `components` prop. If you provide `null`, then that component will be hidden instead.
