--- apps/examples/src/examples/custom-debug-menu/README.md
@@ -1,12 +1,0 @@
----
-title: Custom debug menu
-component: ./CustomDebugMenuExample.tsx
-category: ui
-keywords: [components, ui]
----
-
-You can customize tldraw's debug menu.
-
----
-
-The debug menu contains helpful menu items for debugging the tldraw component. To show the debug menu, turn on debug mode in the preferences. It can be customized by providing a `DebugMenu` component to the `Tldraw` component's `components` prop. If you provide `null`, then that component will be hidden.
