--- apps/examples/src/examples/custom-help-menu/README.md
@@ -1,12 +1,0 @@
----
-title: Custom help menu
-component: ./CustomHelpMenuExample.tsx
-category: ui
-keywords: [components, ui]
----
-
-You can customize tldraw's help menu.
-
----
-
-The help menu contains options for opening the keyboard shortcuts dialog and changing the language. It can be opened by clicking on the help icon in the bottom right of the screen at larger breakpoints. The help menu can be customized by providing a `HelpMenu` component to the `Tldraw` component's `components` prop. If you provide `null`, then that component will be hidden.
