--- apps/examples/src/examples/custom-keyboard-shortcuts-dialog/README.md
@@ -1,12 +1,0 @@
----
-title: Custom keyboard shortcuts dialog
-component: ./CustomKeyboardShortcutsDialogExample.tsx
-category: ui
-keywords: [components, ui]
----
-
-You can customize tldraw's keyboard shortcuts dialog.
-
----
-
-The keyboard shortcuts dialog can be opened via the help menu in the bottom right corner of the tldraw component. It can be customized by providing a `KeyboardShortcutsDialog` component to the `Tldraw` component's `components` prop. If you provide `null`, then that component will be hidden instead.
