--- apps/examples/src/examples/custom-main-menu/README.md
@@ -1,12 +1,0 @@
----
-title: Custom main menu
-component: ./CustomMainMenuExample.tsx
-category: ui
-keywords: [components, ui]
----
-
-You can customize tldraw's main menu.
-
----
-
-The main menu contains important submenus: Edit, Shape, Preferences etc. To open the main menu, click the hamburger icon in the top left corner of the tldraw component. It can be customized by providing a `MainMenu` component to the `Tldraw` component's `components` prop. If you provide `null`, then that component will be hidden.
