+++ apps/examples/src/examples/custom-menus/README.md
@@ -1,0 +1,13 @@
+---
+title: Changing default menus
+component: ./CustomMenusExample.tsx
+category: ui
+priority: 1
+keywords: [menu, panel, actions, style, toolbar, main, page, zoom, navigation, quick]
+---
+
+Customize tldraw's menus.
+
+---
+
+You can use the `components` prop to override tldraw's default menus. You can provide a React component of your own, import our default component and edit it, or return null to hide it completely. This example demonstrates how to do this with every menu in tldraw.
