--- apps/examples/src/examples/custom-navigation-panel/README.md
@@ -1,12 +1,0 @@
----
-title: Custom navigation panel
-component: ./CustomNavigationPanelExample.tsx
-category: ui
-keywords: [components, ui]
----
-
-You can customize tldraw's navigation panel or remove it entirely.
-
----
-
-The navigation panel is in the bottom left of the tldraw component at larger breakpoints. It contains zoom controls and a mini map. It can be customized by providing a `NavigationPanel` component to the `Tldraw` component's `components` prop. If you provide `null`, then that component will be hidden instead.
