--- apps/examples/src/examples/custom-page-menu/README.md
@@ -1,12 +1,0 @@
----
-title: Custom page menu
-component: ./CustomPageMenuExample.tsx
-category: ui
-keywords: [components, ui]
----
-
-You can customize tldraw's page menu, or remove it entirely.
-
----
-
-The page menu contains options for creating and editing pages. To open the page menu, click the page name in the top left of the tldraw component. It can be customized by providing a `PageMenu` component to the `Tldraw` component's `components` prop. If you provide `null`, then that component will be hidden.
