--- apps/examples/src/examples/custom-quick-actions/README.md
@@ -1,12 +1,0 @@
----
-title: Custom quick actions
-component: ./CustomQuickActions.tsx
-category: ui
-keywords: [components, ui]
----
-
-You can customize tldraw's quick actions, a collection of components that appear next to the menu button, or in the toolbar on smaller sizes.
-
----
-
-The quick actions component can be customized by providing a `QuickActionsContent` component to the `Tldraw` component's `components` prop. If you provide `null`, then that component will be hidden.
