--- apps/examples/src/examples/custom-style-panel/README.md
@@ -1,12 +1,0 @@
----
-title: Custom style panel
-component: ./CustomStylePanelExample.tsx
-category: ui
-keywords: [components, ui]
----
-
-You can customize tldraw's style panel or remove it entirely.
-
----
-
-The style panel can be customized by providing a `StylePanel` component to the `Tldraw` component's `components` prop. If you provide `null`, then that component will be hidden instead.
