--- apps/examples/src/examples/custom-toolbar/README.md
@@ -1,12 +1,0 @@
----
-title: Custom toolbar
-component: ./CustomToolbarExample.tsx
-category: ui
-keywords: [components, ui]
----
-
-You can customize tldraw's toolbar or remove it entirely.
-
----
-
-The toolbar can be customized by providing a `Toolbar` component to the `Tldraw` component's `components` prop. If you provide `null`, then that component will be hidden instead.
